requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
watchdog==3.0.0
flask-swagger-ui==4.11.1
apispec==6.3.0
apispec-webframeworks==0.5.2
marshmallow==3.20.1
PyJWT==2.8.0
flask-bcrypt==1.0.1
pytest==7.4.0